Fix issue #1 to handle nested loops correctly, along with refactoring and cleanup.
This adds an instruction buffer (instead of maintaining a pointer into
the file itself) and a stack (to correctly track nested loops) to the
interpreter's runtime system. Code has been significantly refactored as
a result.

Also removed the samples from the repo, as I did not write those
programs and don't want to infringe on any license restrictions.